% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vedic_muhurta.R
\name{jd_to_time}
\alias{jd_to_time}
\title{Convert Julian Day to formatted time string (HH:MM:SS)}
\usage{
jd_to_time(jd)
}
\arguments{
\item{jd}{Julian Day number}
}
\value{
Formatted time string in local time zone
}
\description{
Convert Julian Day to formatted time string (HH:MM:SS)
}
