% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vedic_muhurta.R
\name{vedic_muhurta}
\alias{vedic_muhurta}
\title{Calculate Vedic Muhurtas for a given date and location}
\usage{
vedic_muhurta(jd, place)
}
\arguments{
\item{jd}{Julian Day (JD) number of the given date}

\item{place}{A numeric vector containing latitude, longitude,
and timezone offset}
}
\value{
A data frame containing 30 Muhurtas with their respective
start and end times
}
\description{
This function divides the time between two consecutive sunrises
into 30 equal Muhurtas.
}
\examples{
jd_example <- gregorian_to_jd(8, 9, 2022)
place <- c(15.34, 75.13, 5.5)
vedic_muhurta(jd_example, place)
}
